resume_parser_v1/requirement.txt
alabaster==0.7.12
appdirs==1.4.3
argcomplete==1.10.0
asn1crypto==1.2.0
astroid==2.3.3
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
black==19.10b0
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
cryptography==2.8
dateparser==0.7.2
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
elasticsearch==7.1.0
entrypoints==0.3
extract-msg==0.23.1
filelock==3.0.12
flashtext==2.7
future==0.18.2
glob2==0.7
idna==2.8
imagesize==1.1.0
IMAPClient==2.1.0
importlib-metadata==1.1.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
jeepney==0.4.1
Jinja2==2.10.3
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.2.3
jupyterlab-server==1.0.6
keyring==19.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
lxml==4.4.2
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
mysql-connector==2.2.9
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.19.1
olefile==0.46
pandas==1.1.1
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.6.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psutil==5.8.0
pycparser==2.19
Pygments==2.5.2
pyrsistent==0.15.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2019.3
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==18.1.0
regex==2019.12.9
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.5
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.3
tqdm==4.48.2
traitlets==4.3.3
typed-ast==1.4.0
tzlocal==1.5.1
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
